+++ backend/controllers/roomCategoriesControllers.js
@@ -1,4 +1,5 @@
 const Category = require("../models/roomCategoryModel");
+const Room = require("../models/roomModel");
 
 const getAllCategories = async (req, res) => {
   try {
@@ -12,8 +13,11 @@ const getAllCategories = async (req, res) => {
 const addNewCategory = async (req, res) => {
   const { nameCategory } = req.body;
   try {
-    const newCategory = await Category.create({ name: nameCategory });
-    res.status(200).json(newCategory);
+    if (nameCategory) {
+      console.log(nameCategory);
+      const newCategory = await Category.create({ name: nameCategory });
+      res.status(200).json(newCategory);
+    }
   } catch (error) {
     res.status(400).json({ error: error.message });
   }
@@ -33,8 +37,18 @@ const getThisCategory = async (req, res) => {
 const deleteThisCategory = async (req, res) => {
   const { id } = req.params;
   try {
-    const thisCategory = await Category.findByIdAndDelete({ _id: id });
-    res.status(200).json({ message: "Category deleted successfully" });
+    const roomsByCategory = await Room.find({ category: id });
+    if (roomsByCategory.length > 0) {
+      console.log(roomsByCategory.length);
+      res.status(200).json({
+        message:
+          "You cannot delete this category. The category has associated rooms.",
+      });
+    } else {
+      console.log("No rooms found in this category.");
+      const thisCategory = await Category.findByIdAndDelete({ _id: id });
+      res.status(200).json({ message: "Category deleted successfully" });
+    }
   } catch (error) {
     res.status(400).json({ error: error.message });
   }
